Move Pioneer utils to new package `internal`
