+++ src/main/java/org/junitpioneer/internal/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Package for JUnit Pioneer implementation details.
+ * DO NOT USE THESE CLASSES - THEY MAY CHANGE SIGNIFICANTLY IN ANY MINOR UPDATE.
+ */
+
+package org.junitpioneer.internal;
